requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
elastic-transport==8.12.0
elasticsearch==8.12.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.9.0
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
orjson==3.9.15
packaging==24.0
pillow==10.2.0
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pymongo==4.6.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.3
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.3.1
sentencepiece==0.2.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1+cpu
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
